All is lost.

+++ Assets/Totem Tower Defence/Scripts/UI/EnemyLifeUI.cs
@@ -37,7 +37,7 @@
             }
         }
 
-        public void SetLife ( int value ) {
+        public void SetLife ( float value ) {
             FadeIn();
             slider.value = value;
         }
